Extrae selección de country a clase aparte
