+++ core/src/main/screens/RunningScreen.kt
@@ -1,23 +1,18 @@
 package screens
 
+import CountrySelector
 import Kamchatka
-import MapColors
 import com.badlogic.gdx.Input
-import com.badlogic.gdx.graphics.Color
-import com.badlogic.gdx.graphics.Pixmap
 import com.badlogic.gdx.graphics.Texture
 import com.badlogic.gdx.graphics.g2d.BitmapFont
 import com.badlogic.gdx.graphics.g2d.Sprite
-import com.badlogic.gdx.math.Vector3
 import com.badlogic.gdx.utils.viewport.FitViewport
 
 class RunningScreen(private val game: Kamchatka) : KamchatkaScreen(game) {
     private val worldmap: Sprite = Sprite(Texture("mapa.png"))
-    private val countryColorsMap: Texture = Texture("colores-paises.png")
     private var currentCountry: String? = null
-    private val mapColors = MapColors.fromJsonFile("mapa.json")
     private val message = BitmapFont()
-    private val worldmapPixmap: Pixmap
+    private val countrySelector = CountrySelector(worldmap, game.camera)
 
     init {
         game.viewport = FitViewport(
@@ -25,11 +20,6 @@ class RunningScreen(private val game: Kamchatka) : KamchatkaScreen(game) {
             worldmap.height,
             game.camera
         )
-        val textureData = countryColorsMap.textureData
-        if (!textureData.isPrepared) {
-            textureData.prepare()
-        }
-        worldmapPixmap = textureData.consumePixmap()
         worldmap.setPosition(-worldmap.width / 2, -worldmap.height / 2)
     }
 
@@ -56,44 +46,21 @@ class RunningScreen(private val game: Kamchatka) : KamchatkaScreen(game) {
         worldmap.texture.dispose()
     }
 
-
     override fun touchDown(
         screenX: Int, screenY: Int, pointer: Int, button: Int
     ): Boolean {
-        val color = getMapColorAtPoint(screenX, screenY)
-        if (color in mapColors) {
-            val country = mapColors[color]
-            println("position: ($screenX, $screenY), country: $country, color: $color.")
+        val country = countrySelector.selectByScreenPosition(screenX, screenY)
+        if (country == null) {
+            println("There's no country here.")
         } else {
-            println("There's no country in position ($screenX, $screenY).")
+            println("Selected country $country.")
         }
         return true
     }
 
     override fun mouseMoved(screenX: Int, screenY: Int): Boolean {
-        val color = getMapColorAtPoint(screenX, screenY)
-        currentCountry = mapColors[color]
+        currentCountry = countrySelector.selectByScreenPosition(screenX, screenY)
         return true
-    }
-
-    private fun getMapColorAtPoint(screenX: Int, screenY: Int): Color {
-        val (actualX, actualY) = screenPositionToWorldMapPosition(screenX, screenY)
-        return Color(worldmapPixmap.getPixel(actualX, actualY))
-    }
-
-    private fun screenPositionToWorldMapPosition(
-        screenX: Int, screenY: Int
-    ): Pair<Int, Int> {
-        val screenPosition = Vector3(screenX.toFloat(), screenY.toFloat(), 0F)
-        val worldPosition = game.camera.unproject(screenPosition)
-        // worldPosition starts from the top left,
-        // spritePosition starts from the bottom left.
-        // This means that the y coordinate must be "inverted"
-        return Pair(
-            (worldPosition.x - worldmap.x).toInt(),
-            (worldmap.height - (worldPosition.y - worldmap.y)).toInt()
-        )
-
     }
 
 }
